requirements.txt
streamlit>=1.52.0
google-generativeai>=0.8.6
faiss-cpu>=1.13.0
pypdf>=6.5.0
python-docx>=1.1.0
newspaper3k>=0.2.8
numpy>=1.26.0